Converts npm run pylint command to use js.
Currently `npm run pylint` triggers a shell script to run pylint.
This a) doesn't work on windows b) doesn't use correct path to pylint
because depot_tools isn't in the PATH.

This change adds the same functionality that the shell script provides
to our js scripts. It also adds depot_tools/python2-bin and python2-bin
to the PATH as recommended by Chromium.